news-poller/requirements.txt
kafka-python==2.0.2
feedparser==6.0.11
requests>=2.31.0
python-dotenv==1.0.0
